requirements.txt
flask
beautifulsoup4
requests
python-dotenv